chore: refactor `structure` command, fixes
Refactors the `structure` command to support recursive structures. These are disabled for now, pending additional elaborator support in #5822. This refactor is also a step toward `structure` appearing in `mutual` blocks.

Error reporting is now more precise, and this fixes an issue where general errors could appear on the last field.

Closes #2512