feat: stronger proof strategy for `partial` inhabitation
Now `partial` will try to extract more inhabitants of types by trying to find inhabitants of arguments' arguments.

For example, in a callback like
```lean
partial def withCallback (k : Array Nat → IO α) : IO α := ...
```
it can see that `Array Nat` is inhabited, so with `k` it gets an inhabitant of `IO α`.