feat: record all structure parents in `StructureInfo`
Followup to #5841. Makes the `structure` command populate the new `parentInfo` field.

This will require a stage0 update to fully take effect.

Breaking change: now `getParentStructures` is `getStructureSubobjects`. Adds `getStructureParentInfo` for getting all the immediate parents. Note that the set of subobjects is neither a subset nor a superset of the immediate parents.

Closes #1881