fix: make structure parent info persist
Modifies the `structureExt` from being a `SimplePersistentEnvExtension` to a `PersistentEnvExtension`. The simple version contains a `List` of all added entries, which we do not need since we already have a `PersistentHashMap` of them in the state. The oversight was that this `List` contained duplicate entries due to `setStructureParents` re-adding entries.