feat: make `MapDeclarationExtension` tolerate multiple insertions
Simplifies the definition of `MapDeclarationExtension` so that it only contains a `NameMap` without an additional `List (Name × α)`. Uses the `NameMap`'s natural ordering during export rather than sorting.

This fixes issues from inserting into a `MapDeclarationExtension` multiple times with the same key. Inside a module it appears that each insertion overwrites the data, since those queries access the `NameMap`. But across modules, only the first insertion is accessible, since each insertion was actually pushed to the front of a `List`.

Mathlib needs this for a documentation extension feature, and [they are considering a PR with a workaround](https://github.com/leanprover-community/mathlib4/pull/17043) that digs into the `MapDeclarationExtension` data structures.