fix: `..` in patterns should not make use of opt-params
In patterns, ellipsis should always be an implicit argument, even if it is an optparam or autoparam. This prevents examples such as the one in #4555 from failing:
```lean
match e with
| .internal .. => sorry
| .error .. => sorry
```
The `internal` constructor has an optparam (`| internal (id : InternalExceptionId) (extra : KVMap := {})`).

We may consider having ellipsis suppress optparams and autoparams in general. We avoid doing so for now since it's possible to opt-out of them individually (for example with `.internal (extra := _) ..`) but it's not possible to opt-in. While we can opt-out like this for patterns as well, with patterns it is suprising how strongly the optparams are inserted.

Closes #4555