feat: simp -underLambda
This adds a new simp configuration flag: `underLambda`, on by default.

Turning this off will make `simp` not rewrite under lambdas (and not in
the branches of an `if-then-else`). This can can be useful when
one wants to simplify in a way that more closely matches call-by-value
execution, and can help with `simp [f]` where `f` is a function defined
by well-founded recursion, and where without the flag the simplifier
would go into a loop or produce huge proof terms.

Preventing simplification under lambda is straight-forward (in
`simpLambda`. Recognizing all the places where we might descent into the
branches of if-then-else is a bit more work.