refactor: omega: avoid MVar machinery
@TwoFX mentioned that in his work on weight-balanced trees he has some
`omega` proofs that take very long (16s), had many disjunctions, and
already traced it down to `instantiateMVars`. So I got curious and had a
look how omega is using the `MVar` machinery, and it is only using it to
do case analysis on `Or`.

It wasn't hard to rewrite this code to to construct the proof object
directly, without using `MVars`, using `Or.elim`, and indeed it brings
down the processing time of that particular proof from 16s to just below
one second. And we even get a net reduction of the line count!