feat: omega to abstract its own proofs
This PR is an experiment.

Some experiments show that abstracting the proofs produced by omega can
speed up a particularly gnarly proof from 15s to 10s. Not a
game-changer, but still notable.

It seems this breaks the use of `omega` as a simp discharger. Are simp
dischargers not allowed to write auxillary definitions to the
environment?