feat: warnings when declarations and `export` aliases conflict
This PR adds warnings when a declaration is added that could conflict with a pre-existing `export` alias or vice versa. The warning can be disabled with `set_option linter.aliasConflict false`. The checking is *not* done when importing modules, so there could still be un-warned conflicts between aliases and declarations in diamond imports.

Closes #5258